requirements.txt
# Charlie application requirements
# Core dependencies
requests>=2.28.1
pyyaml>=6.0
python-dotenv>=1.0.0
psutil>=5.9.0

# Web UI
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
beautifulsoup4>=4.12.0
ollama
TTS>=0.21.0


# Voice assistant integration
git+https://github.com/mezbaul-h/june.git
transformers>=4.35.0
torch>=2.0.0
numpy>=1.22.0
sounddevice>=0.4.5
soundfile>=0.12.1
pytest>=8.0.0
pyttsx3>=2.90
py>=1.11.0
